+++ tests/testthat/test-stantipred.R
@@ -48,7 +48,7 @@ if(identical(Sys.getenv("NOT_CRAN"), "true")& .Machine$sizeof.pointer != 4){
     
     checkm$pars[c(-1,-7) ,c('TI1_effect')] <- FALSE
     
-    tfit1<-ctStanFit(tdat,checkm,chains=1,cores=2,optimize=TRUE,
+    tfit1<-ctStanFit(tdat,checkm,cores=2,optimize=TRUE,
       optimcontrol=list(is=TRUE,finishsamples=500),
       priors=TRUE,verbose=0)
     s1=summary(tfit1)